moving folder app "users" in app/services/auth
