+++ app/services/auth/users/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class UsersConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'app.users'
+    name = 'app.services.auth.users'
