+++ app/services/auth/users/logic/mail.py
@@ -7,7 +7,7 @@ from smtplib import (
 )
 from django.core.mail import send_mail
 
-from app.users.exceptions import SendEmailError
+from app.services.auth.users.exceptions import SendEmailError
 
 logger = logging.getLogger(__name__)
